refactor: views tests
